backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
groq==0.4.2
python-jose[cryptography]==3.3.0
faiss-cpu==1.7.4
numpy==1.26.4
langchain-huggingface==0.0.3
sentence-transformers==2.5.1
PyMuPDF==1.23.21
arxiv==2.1.0
